pyproject.toml
[tool.poetry]
name = "lesson-27-hw-django"
version = "0.1.0"
description = ""
authors = ["VladmimirShtefan <shtefanvladimir14@gmail.com>"]
readme = "README.md"
packages = [{include = "lesson_27_hw_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
